# scripts/openapi_schema.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Cadence v2",
    "description": "\n# Multi-Tenant AI Orchestration Platform\n\nProduction-ready framework for deploying AI agent orchestrators at scale.\n\n## Features\n\n* **Multi-Backend Support**: LangGraph, OpenAI Agents SDK, Google ADK\n* **Multi-Tenancy**: Complete tenant isolation with BYOK (Bring Your Own Key)\n* **Three Orchestration Modes**: Supervisor, Coordinator, Handoff\n* **Hot-Reload**: Configuration changes without restart\n* **Scalable Architecture**: Hot/Warm/Cold tier pool (1000+ instances)\n* **Streaming**: Server-Sent Events (SSE) for real-time responses\n\n## Authentication\n\n### JWT Authentication\n```\nAuthorization: Bearer <jwt_token>\n```\n\n### API Key Authentication\n```\nX-API-Key: <api_key>\n```\n\n## Rate Limiting\n\nAll API endpoints are rate-limited per tenant. Default: 100 requests/minute.\n\n## Documentation\n\n* Platform Docs: [README](https://github.com/jonaskahn/cadence)\n* SDK Docs: [cadence-sdk](https://github.com/jonaskahn/cadence-sdk)\n",
    "version": "2.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Local development server"
    },
    {
      "url": "https://api.example.com",
      "description": "Production server"
    },
    {
      "url": "https://staging-api.example.com",
      "description": "Staging server"
    }
  ],
  "paths": {
...
    "/api/orgs/{org_id}/plugins": {
      "get": {
        "tags": [
          "plugins"
        ],
        "summary": "List Plugins",
        "description": "List all available plugins for tenant (system + org combined).\n\nArgs:\n    request: FastAPI request\n    tag: Optional tag filter\n    context: Tenant context from JWT\n\nReturns:\n    List of plugin metadata",
        "operationId": "list_plugins_api_orgs__org_id__plugins_get",
        "parameters": [
          {
            "name": "tag",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tag"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PluginMetadataResponse"
                  },
                  "title": "Response List Plugins Api Orgs  Org Id  Plugins Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/orgs/{org_id}/plugins/upload": {
      "post": {
        "tags": [
          "plugins"
        ],
        "summary": "Upload Org Plugin",
        "description": "Upload an org-specific plugin package.\n\nArgs:\n    file: Plugin zip file\n    request: FastAPI request\n    context: Tenant context from JWT\n\nReturns:\n    Upload confirmation with pid and version",
        "operationId": "upload_org_plugin_api_orgs__org_id__plugins_upload_post",
        "parameters": [
          {
            "name": "org_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Org Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_org_plugin_api_orgs__org_id__plugins_upload_post"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/orgs/{org_id}/plugins/{plugin_id}": {
      "delete": {
        "tags": [
          "plugins"
        ],
        "summary": "Delete Org Plugin",
        "description": "Soft-delete an org plugin.\n\nArgs:\n    plugin_id: Plugin database ID\n    request: FastAPI request\n    context: Tenant context from JWT",
        "operationId": "delete_org_plugin_api_orgs__org_id__plugins__plugin_id__delete",
        "parameters": [
          {
            "name": "plugin_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Plugin Id"
            }
          },
          {
            "name": "org_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Org Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/plugins": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "List System Plugins",
        "description": "List all system plugins, optionally filtered by tag.\n\nArgs:\n    request: FastAPI request\n    tag: Optional tag filter\n    tenant_context: Tenant context from JWT\n\nReturns:\n    List of system plugin metadata",
        "operationId": "list_system_plugins_api_admin_plugins_get",
        "parameters": [
          {
            "name": "tag",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tag"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SystemPluginResponse"
                  },
                  "title": "Response List System Plugins Api Admin Plugins Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/plugins/upload": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Upload System Plugin",
        "description": "Upload a system plugin package.\n\nArgs:\n    file: Plugin zip file\n    request: FastAPI request\n    tenant_context: Tenant context from JWT\n\nReturns:\n    Upload confirmation with pid and version",
        "operationId": "upload_system_plugin_api_admin_plugins_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_system_plugin_api_admin_plugins_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/plugins/{plugin_id}": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Get System Plugin",
        "description": "Get a system plugin by ID.\n\nArgs:\n    plugin_id: Plugin database ID\n    request: FastAPI request\n    tenant_context: Tenant context from JWT\n\nReturns:\n    System plugin details",
        "operationId": "get_system_plugin_api_admin_plugins__plugin_id__get",
        "parameters": [
          {
            "name": "plugin_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Plugin Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SystemPluginResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "admin"
        ],
        "summary": "Soft Delete System Plugin",
        "description": "Soft-delete a system plugin (sets is_active=False).\n\nArgs:\n    plugin_id: Plugin database ID\n    request: FastAPI request\n    tenant_context: Tenant context from JWT",
        "operationId": "soft_delete_system_plugin_api_admin_plugins__plugin_id__delete",
        "parameters": [
          {
            "name": "plugin_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Plugin Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_upload_org_plugin_api_orgs__org_id__plugins_upload_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_org_plugin_api_orgs__org_id__plugins_upload_post"
      },
      "Body_upload_system_plugin_api_admin_plugins_upload_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_system_plugin_api_admin_plugins_upload_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PluginMetadataResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "pid": {
            "type": "string",
            "title": "Pid"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "version": {
            "type": "string",
            "title": "Version"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "tag": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tag"
          },
          "is_latest": {
            "type": "boolean",
            "title": "Is Latest"
          },
          "capabilities": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Capabilities"
          },
          "agent_type": {
            "type": "string",
            "title": "Agent Type"
          },
          "stateless": {
            "type": "boolean",
            "title": "Stateless"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "default_settings": {
            "additionalProperties": true,
            "type": "object",
            "title": "Default Settings",
            "default": {}
          }
        },
        "type": "object",
        "required": [
          "id",
          "pid",
          "name",
          "version",
          "description",
          "is_latest",
          "capabilities",
          "agent_type",
          "stateless",
          "source"
        ],
        "title": "PluginMetadataResponse",
        "description": "Plugin metadata response.\n\nAttributes:\n    id: Database primary key\n    pid: Reverse-domain plugin identifier\n    name: Plugin name\n    version: Plugin version\n    description: Plugin description\n    tag: Free-form search tag\n    is_latest: Whether this is the latest version\n    capabilities: List of capabilities\n    agent_type: Agent type (specialized/general)\n    stateless: Whether plugin is stateless\n    source: Plugin source (system/org)\n    default_settings: Default settings keyed by field name"
      },
...
      "SystemPluginResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "pid": {
            "type": "string",
            "title": "Pid"
          },
          "version": {
            "type": "string",
            "title": "Version"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "tag": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tag"
          },
          "is_latest": {
            "type": "boolean",
            "title": "Is Latest"
          },
          "s3_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "S3 Path"
          },
          "default_settings": {
            "additionalProperties": true,
            "type": "object",
            "title": "Default Settings",
            "default": {}
          },
          "capabilities": {
            "items": {},
            "type": "array",
            "title": "Capabilities",
            "default": []
          },
          "agent_type": {
            "type": "string",
            "title": "Agent Type"
          },
          "stateless": {
            "type": "boolean",
            "title": "Stateless"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          }
        },
        "type": "object",
        "required": [
          "id",
          "pid",
          "version",
          "name",
          "is_latest",
          "agent_type",
          "stateless",
          "is_active"
        ],
        "title": "SystemPluginResponse",
        "description": "System plugin response model."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtained from authentication endpoint. Format: `Bearer <token>`"
      },
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API key for service-to-service authentication. Format: `X-API-Key: <key>`"
      }
    }
  },
...
  "security": [
    {
      "BearerAuth": []
    },
    {
      "ApiKeyAuth": []
    }
  ],
...
}